Copied contents of LineChart.jsx to LineChart-fix.jsx
